Loader.js was modified. Added server.js , billingCycle.js and database.js
