+++ CursoFramesWeb/backend/api/billingCycle/billingCycle.js
@@ -1,0 +1,26 @@
+const restful = require('node-restful')
+const mongoose = restful.mongoose
+
+//Definindo um esquema para o crédito
+const creditSchema = new mongoose.Schema({
+	name: {type: String, required : true},
+	value: { type: Number, min: 0, required: true}
+})
+
+//Definindo um esquema para o débito
+const debtSchema = new mongoose.Schema({
+	name: {type: String, required : true},
+	value: { type: Number, min: 0, required: true},
+	status: { type: String, required: false, uppercase: true,
+		enum: ['PAGO','PENDENTE','AGENDADO']}
+})
+
+const billingCycleSchema = new mongoose.Schema({
+	name: { type: String, required: true},
+	month: { type: Number, min: 1 ,max: 12, required: true},
+	year: { type: Number, min: 1970, max: 2100, required: true},
+	credits : [creditSchema],
+	debits: [debtSchema]
+})
+
+module.exports = restful.model('BillingCycle', billingCycleSchema)
