+++ CursoFramesWeb/backend/config/database.js
@@ -1,0 +1,2 @@
+const mongoose = require('mongoose')
+module.exports = mongoose.connect('mongodb://localhost/db_finance')
